refactor: создана папка pages для страниц, содержимое компонента App перемещено в компонент pages/Home
