--- src/components/app/app.module.scss
@@ -1,59 +1,0 @@
-@use "../../styles/variables.scss" as *;
-
-$headerHeight: 90px;
-$mainTitleHeight: 60px;
-$mainTopSpacer: $spacer * 10;
-
-.app {
-  &__container {
-    @include media-bp-up(desktop) {
-      display: grid;
-      grid-template-columns: 1fr 1fr;
-      grid-template-rows: min-content min-content;
-      column-gap: 40px;
-    }
-  }
-
-  &__title {
-    @include media-bp-up(desktop) {
-      grid-area: 1 / 1 / span 1 / -1;
-    }
-
-    @include media-bp-down(desktop) {
-      max-width: 600px;
-      margin-inline: auto;
-    }
-  }
-
-  &__ingredients,
-  &__constructor {
-    width: 100%;
-    max-width: 600px;
-    margin-inline: auto;
-
-    @include media-bp-up(desktop) {
-      display: flex;
-      flex-direction: column;
-      height: calc(100vh - #{$headerHeight} - #{$mainTopSpacer} - #{$mainTitleHeight} - 30px);
-      max-height: calc(100vh - #{$headerHeight} - #{$mainTopSpacer} - #{$mainTitleHeight} - 30px);
-    }
-  }
-
-  &__ingredients {
-    @include media-bp-up(desktop) {
-      grid-area: 2 / 1 / span 1 / span 1;
-    }
-
-    @include media-bp-down(desktop) {
-      margin-bottom: $spacer * 15;
-    }
-  }
-
-  &__constructor {
-    width: 100%;
-
-    @include media-bp-up(desktop) {
-      grid-area: 2 / 2 / span 1 / span 1;
-    }
-  }
-}
