+++ src/pages/home/home.module.scss
@@ -1,0 +1,57 @@
+@use "../../styles/variables.scss" as *;
+
+$headerHeight: 90px;
+$mainTitleHeight: 60px;
+$mainTopSpacer: $spacer * 10;
+
+.container {
+  @include media-bp-up(desktop) {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-rows: min-content min-content;
+    column-gap: 40px;
+  }
+}
+
+.title {
+  @include media-bp-up(desktop) {
+    grid-area: 1 / 1 / span 1 / -1;
+  }
+
+  @include media-bp-down(desktop) {
+    max-width: 600px;
+    margin-inline: auto;
+  }
+}
+
+.ingredients,
+.burger-constructor {
+  width: 100%;
+  max-width: 600px;
+  margin-inline: auto;
+
+  @include media-bp-up(desktop) {
+    display: flex;
+    flex-direction: column;
+    height: calc(100vh - #{$headerHeight} - #{$mainTopSpacer} - #{$mainTitleHeight} - 30px);
+    max-height: calc(100vh - #{$headerHeight} - #{$mainTopSpacer} - #{$mainTitleHeight} - 30px);
+  }
+}
+
+.ingredients {
+  @include media-bp-up(desktop) {
+    grid-area: 2 / 1 / span 1 / span 1;
+  }
+
+  @include media-bp-down(desktop) {
+    margin-bottom: $spacer * 15;
+  }
+}
+
+.burger-constructor {
+  width: 100%;
+
+  @include media-bp-up(desktop) {
+    grid-area: 2 / 2 / span 1 / span 1;
+  }
+}
